fixed some review problems
