+++ package.json
@@ -50,6 +50,7 @@
     "@types/jest": "^27.0.2",
     "@types/node": "^16.11.4",
     "@types/react": "^17.0.32",
+    "@types/react-dom": "^17.0.11",
     "@types/testing-library__jest-dom": "^5.14.1",
     "@types/webpack": "^5.28.0",
     "@typescript-eslint/eslint-plugin": "^5.1.0",
@@ -80,7 +81,6 @@
     "webpack-dev-server": "^4.3.1"
   },
   "dependencies": {
-    "@types/react-dom": "^17.0.10",
     "modern-normalize": "^1.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
