+++ src/lesson3/components/Cell.tsx
@@ -1,4 +1,6 @@
 import React from "react";
+import { useSelector } from "react-redux";
+import { IState } from "../redux/store";
 
 export interface ICell {
   number: number;
@@ -6,7 +8,11 @@ export interface ICell {
   alive: boolean;
 }
 
-type CellProps = ICell;
+export interface ICellProps {
+  number: number;
+  width: number;
+  onCellClick: (number: number) => void;
+}
 
 export function isCell(cell: ICell | unknown): cell is ICell {
   if (cell === null || typeof cell !== "object") {
@@ -20,16 +26,39 @@ export function isCell(cell: ICell | unknown): cell is ICell {
   );
 }
 
-const Cell = (props: CellProps) => {
-  const { number, backgroundColor, alive } = props;
+const Cell = (props: ICellProps) => {
+  const { number, onCellClick, width } = props;
+  const cell: ICell | undefined = useSelector((state: IState) =>
+    state.cells.find((c) => {
+      return c.number === number;
+    })
+  );
+
+  if (!isCell(cell)) {
+    return null;
+  }
+
+  const cellContainerStyles = {
+    width,
+    height: width,
+    fontSize: width,
+  };
 
   const cellStyles: object = {
-    backgroundColor: alive ? backgroundColor : null,
+    backgroundColor: cell.alive ? cell.backgroundColor : null,
   };
 
   return (
-    <div className="cell" style={cellStyles} data-testid="cell">
-      {alive ? number : null}
+    <div
+      className="container-cell"
+      style={cellContainerStyles}
+      onClick={() => {
+        onCellClick(number);
+      }}
+    >
+      <div className="cell" style={cellStyles} data-testid="cell">
+        {cell.alive ? number : null}
+      </div>
     </div>
   );
 };
